Making git integration more submodule friendly
Former-commit-id: 3cd849a7d21e1cac0d9d8f938aa3b9b9ca5cbe20
Former-commit-id: a6b2c002d2ebc8cba49ff91cd67e26fc7ea449ce
Former-commit-id: 236710c8e6b1cd825138e39794c5ba02b8db50fc
Former-commit-id: c268b48341f544d76ad31eca310d7b89462c50e0
Former-commit-id: ae997a7f056e5dfd854cdaf2eab638ba64e8d5a8
Former-commit-id: bda30fedf1d14756c6203c0d65fdb3d74543a5d1
Former-commit-id: 4d2cb1806f4302f50af3a0030a6a397d6fe3dcc8
Former-commit-id: e4bd0355678b27d14bd8f439783025b15efb6465

+++ PowerShellConsole/Exports/Functions/GitHub/_Internal/_invokeGitCommand.ps1
@@ -18,24 +18,32 @@ function _invokeGitCommand {
 		$arguments += @{ "RedirectStandardOutput" = $standardOutput }
 		$standardError = [System.IO.Path]::GetTempFileName()
 		$arguments += @{ "RedirectStandardError" = $standardError }
-		
+
+		$cleanup = {
+			if (Test-Path $standardOutput) { Remove-Item $standardOutput -Force }
+			if (Test-Path $standardError) { Remove-Item $standardError -Force }
+		}		
+	} else {
+		$cleanup = {}
 	}
 
     $gitProcess = Start-Process @arguments
-
-    if ($Quiet -and (Test-Path $standardOutput)) {
-    	Remove-Item $standardOutput -Force
-    }
     
     if ($gitProcess.ExitCode -ne 0 -and $null -ne $gitProcess.ExitCode) {
-    	if ($NonFatalError) { return $gitProcess.ExitCode }
+    	if ($NonFatalError) {
+    		& $cleanup
+    		return $gitProcess.ExitCode
+    	}
     	else {
     		$errorMessage = "Git command ($Command) returned exit code: $($gitProcess.ExitCode)"
     		if ($Quiet) {
     			$errorMessage += "`nCommand Output: $(Get-Content $standardOutput)"
     			$errorMessage += "`nError Message: $(Get-Content $standardError)"
     		}
+    		& $cleanup
     		throw $errorMessage
     	}
+    } else {
+		& $cleanup
     }
 }
